Added projectileSpeed backend logic to Shroomerang.
