+++ Assets/Game/Production/Enemies/BossEnemy/BossEnemy.cs
@@ -1,7 +1,9 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+[Obsolete(" This no longer exists in scope", true)]
 public class BossEnemy : Enemy
 {
     /*
