+++ Assets/Game/Production/Utility/UI/OpenPauseUI.cs
@@ -1,5 +1,6 @@
 using System.Collections;
 using System.Collections.Generic;
+using UnityEditor;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -26,10 +27,10 @@ public class OpenPauseUI : MonoBehaviour
     {
         Debug.Log("Quit button has ben pressed but no unfortunately you can't quit unity.");
         Debug.Log("I meant the editor...");
-        #if UNITY_EDITOR
-                UnityEditor.EditorApplication.isPlaying = false;
+#if UNITY_EDITOR
+        EditorApplication.Exit(0);
         #else
                     Application.Quit();
-        #endif
+#endif
     }
 }
