--- Assets/Game/Production/Utility/UI/QuitScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 169837d35369eff4b969bb7ac11fba07
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
